Support multi-Assert and multi-Retract of same statement
This is useful for streaming -- you can retract all your old
statements without worrying that you'll stomp any new ones that happen
to be the same as the old ones. Each null-match parent of a
statement (representing that statement being asserted) is mapped to a
refcount now, instead of just the `true` sentinel.