Refactor sharing to have a more general 'peers' system.
This is in prep for interprocess communication. For now, it is just
used to implement pretty much the same generational
laptop-sharing-to-table scheme, although done with explicit
share-patterns instead of implicitly based on assertion, which is
nice.

Also make `Claim` and `Wish` store the calling $this instead of
`someone`.

Also start on `Do` for controlling side effects.

Also move back to the old way of collecting matches & build it into
main -- this is used as a hack for negation (so we can explicitly
retract the Pi root statements when a laptop connects)