Revert most of the join implementation in favor of a desugar impl
Joins won't work with `Assert when` now, but they will work with
ordinary `When`.

The motivation is to make it possible to use `When when` with every
clause of a join, so every clause has to exist as a separate When in
the database. (Another point is that the old join implementation was a
significant jump in core complexity without an accompanying
performance increase, so I don't really want to keep it around.)