process: Allow running multiple On process blocks on process
Turn off share wishes/claims default for subprocesses, so you (for
now) have to explicitly specify sharing if you want any to happen.

On process now only spawns process if it doesn't exist but always runs
the code block (which now happens through a statement instead of
directly on init).