Add ws.commit and use for web editor region to fix leak
Make WebSocket receiver do a bare eval and then do the assert wrap on
the ws.send js side. (We might want to rename ws.send to ws.run or
ws.say or something to reflect that it isn't a bare send.)

Doesn't fix closing tab -> program doesn't disappear, but that wasn't
working on this branch already anyway.