Fix esc-restart, remove sudo from keymap
Instead of using `loadkeys` we're now using `kbd_mode -u` to load the
keymap. This is because `loadkeys` requires root permissions and we're
trying to remove the need for `sudo` in the system. I tried removing
`sudo` from the `esc-restart` virtual program but it failed with an
interactive authentication error. I'm not sure how to do this with less
powerful permissions than `sudo` so I'm leaving it as is for now.