db: WIP: Redo statement and match synchronization & abort
Mostly works again but crashes on collision. Two big changes: 1. we
null out the child lists in matches and statements before we clear
them, so that when you insert a child, you're guaranteed that it will
actually get cleared. 2. we reorganize insertion logic so any failure
aborts the whole match/statement creation (?)