WIP: Do recollect immediately, at C level
This may be somewhat slower than before, because it doesn't
deduplicate/debounce recollects at all (what if you do a bunch of
statement adds/removes in succession? we'll have to recollect for
every single one). But it may also be faster, because there should be
much less overhead per recollect (we don't have to pingpong through
the timestamped Hold with the triggering statement, then match that
with When, then eval...)