WIP: Make keepMs actually keep the statement in-index as well
Before this, we were just keeping the descendants in-index, which was
simpler to implement but made it hard to keep writable-images (since
those are direct children of the create-writable-image
statement). We'll see if this works. This does feel more
correct (although idk if the implementation is fully correct yet).

Moving deindex into statementRemoveSelf caused annoying issue where
transient statements got removed -> deindexed -> original statement
vanished, so added deindex param to statementRemoveSelf.