WIP: State camera frame as compressed JPEG; use in /camera-frame
Instead of putting the JPEG data in the db as I was planning to, just
using another fat pointer / slice (like we do with Image) for
now. This is to avoid risk of lots of unnecessary in/out copies in
Tcl.

Haven't implemented any logic to automatically decompress yet.

Nice bonus: /camera-frame now gets to show color as well.